chore(communities)_: mv `unmarshalCommunityDescriptionMessage` to manager
